requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
pandas
numpy
faiss-cpu
sentence-transformers
kafka-python
streamlit
pyarrow